refactor: stream processing
