+++ apps/web/src/lib/constants.ts
@@ -5,3 +5,12 @@ export const MM_CARD_ADDRESSES = [
     // International 
     '0x9dd23A4a0845f10d65D293776B792af1131c7B30'
 ]
+
+// https://developers.circle.com/stablecoins/usdc-contract-addresses
+export const USDC_ADDRESSES = [
+    // Linea
+    '0x176211869cA2b568f2A7D4EE941E073a821EE1ff',
+
+    // Linea Sepolia
+    '0xFEce4462D57bD51A6A552365A011b95f0E16d9B7'
+]
